move offset publisher to types to share with client
