+++ src/spu/Cargo.toml
@@ -39,7 +39,7 @@ event-listener = "2.4.0"
 async-io = "1.3.1"
 
 # Fluvio dependencies
-fluvio-types = { version = "0.2.0", features = ["events"], path = "../types" }
+fluvio-types = { version = "0.2.1", features = ["events"], path = "../types" }
 fluvio-storage = { version = "0.3.0", path = "../storage" }
 fluvio-controlplane = { version = "0.6.0", path = "../controlplane" }
 fluvio-controlplane-metadata = { version = "0.6.0", path = "../controlplane-metadata" }
