+++ src/types/Cargo.toml
@@ -13,3 +13,7 @@ events = ["event-listener"]
 [dependencies]
 tracing = "0.1.19"
 event-listener = { version = "2.5.1", optional = true }
+
+[dev-dependencies]
+fluvio-future = { version = "0.1.0", features = ["fixture", "subscriber"] }
+tokio = { version = "0.2.21", features = ["macros"] }
